add .editorconfig file for basic formatting
EditorConfig is a specification to define the most basic code formatting
stuff, and it is supported by many editors and IDEs, either directly or via
plugins, including VSCode/VSCodium, Vim, emacs and more.

It allows to define formatting style related to indentation, charset, end
of lines and trailing whitespaces. It also allows to apply different
formats for different files based on wildcards.

Add the .editorconfig file at the root of the project with a broad generic
configuration for all file types. Then add exceptions for the file types
which follow different conventions.

Signed-off-by: Vincent Mailhol <mailhol.vincent@wanadoo.fr>